moved into SatelliteSoilMoisture folder
